Rename MacOperatingSystem -> MacOperatingSystemJNA
