+++ oshi-core-java25/src/main/java/oshi/software/os/mac/MacOperatingSystemFFM.java
@@ -47,7 +47,7 @@ import oshi.util.tuples.Pair;
  * marketed by Apple Inc. since 2001. It is the primary operating system for Apple's Mac computers.
  */
 @ThreadSafe
-public class MacOperatingSystemFFM extends AbstractMacOperatingSystem {
+public class MacOperatingSystemFFM extends MacOperatingSystem {
 
     private static final Logger LOG = LoggerFactory.getLogger(MacOperatingSystemFFM.class);
