--- oshi-core/src/main/java/oshi/software/os/mac/AbstractMacOperatingSystem.java
@@ -1,179 +1,0 @@
-package oshi.software.os.mac;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import oshi.annotation.concurrent.ThreadSafe;
-import oshi.software.common.AbstractOperatingSystem;
-import oshi.software.os.ApplicationInfo;
-import oshi.software.os.OSProcess;
-import oshi.software.os.OSService;
-import oshi.software.os.OSThread;
-import oshi.util.ExecutingCommand;
-import oshi.util.FileUtil;
-import oshi.util.Memoizer;
-import oshi.util.ParseUtil;
-import oshi.util.Util;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Properties;
-import java.util.Set;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-import static oshi.software.os.OSService.State.RUNNING;
-import static oshi.software.os.OSService.State.STOPPED;
-import static oshi.util.Memoizer.installedAppsExpiration;
-
-/**
- * macOS, previously Mac OS X and later OS X) is a series of proprietary graphical operating systems developed and
- * marketed by Apple Inc. since 2001. It is the primary operating system for Apple's Mac computers.
- */
-@ThreadSafe
-public abstract class AbstractMacOperatingSystem extends AbstractOperatingSystem {
-
-    private static final Logger LOG = LoggerFactory.getLogger(MacOperatingSystem.class);
-
-    public static final String MACOS_VERSIONS_PROPERTIES = "oshi.macos.versions.properties";
-
-    private static final String SYSTEM_LIBRARY_LAUNCH_AGENTS = "/System/Library/LaunchAgents";
-    private static final String SYSTEM_LIBRARY_LAUNCH_DAEMONS = "/System/Library/LaunchDaemons";
-
-    protected final int maxProc;
-
-    protected final String osXVersion;
-    protected final int major;
-    protected final int minor;
-
-    private final Supplier<List<ApplicationInfo>> installedAppsSupplier = Memoizer
-        .memoize(MacInstalledApps::queryInstalledApps, installedAppsExpiration());
-
-    protected AbstractMacOperatingSystem(int maxproc) {
-        String version = System.getProperty("os.version");
-        int verMajor = ParseUtil.getFirstIntValue(version);
-        int verMinor = ParseUtil.getNthIntValue(version, 2);
-        // Big Sur (11.x) may return 10.16
-        if (verMajor == 10 && verMinor > 15) {
-            String swVers = ExecutingCommand.getFirstAnswer("sw_vers -productVersion");
-            if (!swVers.isEmpty()) {
-                version = swVers;
-            }
-            verMajor = ParseUtil.getFirstIntValue(version);
-            verMinor = ParseUtil.getNthIntValue(version, 2);
-        }
-        this.osXVersion = version;
-        this.major = verMajor;
-        this.minor = verMinor;
-        // Set max processes
-        this.maxProc = maxproc;
-    }
-
-    @Override
-    public String queryManufacturer() {
-        return "Apple";
-    }
-
-    protected String parseCodeName() {
-        Properties verProps = FileUtil.readPropertiesFromFilename(MACOS_VERSIONS_PROPERTIES);
-        String codeName = null;
-        if (this.major > 10) {
-            codeName = verProps.getProperty(Integer.toString(this.major));
-        } else if (this.major == 10) {
-            codeName = verProps.getProperty(this.major + "." + this.minor);
-        }
-        if (Util.isBlank(codeName)) {
-            LOG.warn("Unable to parse version {}.{} to a codename.", this.major, this.minor);
-        }
-        return codeName;
-    }
-
-    @Override
-    protected int queryBitness(int jvmBitness) {
-        if (jvmBitness == 64 || (this.major == 10 && this.minor > 6)) {
-            return 64;
-        }
-        return ParseUtil.parseIntOrDefault(ExecutingCommand.getFirstAnswer("getconf LONG_BIT"), 32);
-    }
-
-    @Override
-    public List<OSProcess> queryChildProcesses(int parentPid) {
-        List<OSProcess> allProcs = queryAllProcesses();
-        Set<Integer> descendantPids = getChildrenOrDescendants(allProcs, parentPid, false);
-        return allProcs.stream().filter(p -> descendantPids.contains(p.getProcessID())).collect(Collectors.toList());
-    }
-
-    @Override
-    public List<OSProcess> queryDescendantProcesses(int parentPid) {
-        List<OSProcess> allProcs = queryAllProcesses();
-        Set<Integer> descendantPids = getChildrenOrDescendants(allProcs, parentPid, true);
-        return allProcs.stream().filter(p -> descendantPids.contains(p.getProcessID())).collect(Collectors.toList());
-    }
-
-    @Override
-    public int getThreadId() {
-        OSThread thread = getCurrentThread();
-        if (thread == null) {
-            return 0;
-        }
-        return thread.getThreadId();
-    }
-
-    @Override
-    public OSThread getCurrentThread() {
-        // Get oldest thread
-        return getCurrentProcess().getThreadDetails().stream().sorted(Comparator.comparingLong(OSThread::getStartTime))
-            .findFirst().orElse(new MacOSThread(getProcessId()));
-    }
-
-    @Override
-    public long getSystemUptime() {
-        return System.currentTimeMillis() / 1000 - getSystemBootTime();
-    }
-
-    @Override
-    public List<OSService> getServices() {
-        // Get running services
-        List<OSService> services = new ArrayList<>();
-        Set<String> running = new HashSet<>();
-        for (OSProcess p : getChildProcesses(1, ProcessFiltering.ALL_PROCESSES, ProcessSorting.PID_ASC, 0)) {
-            OSService s = new OSService(p.getName(), p.getProcessID(), RUNNING);
-            services.add(s);
-            running.add(p.getName());
-        }
-        // Get Directories for stopped services
-        ArrayList<File> files = new ArrayList<>();
-        File dir = new File(SYSTEM_LIBRARY_LAUNCH_AGENTS);
-        if (dir.exists() && dir.isDirectory()) {
-            files.addAll(Arrays.asList(dir.listFiles((f, name) -> name.toLowerCase(Locale.ROOT).endsWith(".plist"))));
-        } else {
-            LOG.error("Directory: /System/Library/LaunchAgents does not exist");
-        }
-        dir = new File(SYSTEM_LIBRARY_LAUNCH_DAEMONS);
-        if (dir.exists() && dir.isDirectory()) {
-            files.addAll(Arrays.asList(dir.listFiles((f, name) -> name.toLowerCase(Locale.ROOT).endsWith(".plist"))));
-        } else {
-            LOG.error("Directory: /System/Library/LaunchDaemons does not exist");
-        }
-        for (File f : files) {
-            // remove .plist extension
-            String name = f.getName().substring(0, f.getName().length() - 6);
-            int index = name.lastIndexOf('.');
-            String shortName = (index < 0 || index > name.length() - 2) ? name : name.substring(index + 1);
-            if (!running.contains(name) && !running.contains(shortName)) {
-                OSService s = new OSService(name, 0, STOPPED);
-                services.add(s);
-            }
-        }
-        return services;
-    }
-
-    @Override
-    public List<ApplicationInfo> getInstalledApplications() {
-        return installedAppsSupplier.get();
-    }
-}
